collision update
Pieces now Land properly and pile upwards/
Piece only move left/right/down if they will stay on the board AND if
there is no other piece in the way.

known bugs:
- rotation can cause piece to leave board bounds
- after a piece has landed and before it locks into place, user may
slide it sideways off of the landed-block, but the piece still 'locks
into place'.  Need to catch this un-landed case.